extracted workers from userscontroller, cleaned up businessescontroller
